move fetchAgencyOptions out of admin folder
