+++ front/src/core-logic/domain/admin/adminPreloadedState.ts
@@ -1,9 +1,9 @@
-import { agencyAdminInitialState } from "src/core-logic/domain/admin/agenciesAdmin/fetch-agency-options/fetchAgencyOptions.slice";
 import { connectedUsersAdminInitialState } from "src/core-logic/domain/admin/connectedUsersAdmin/connectedUsersAdmin.slice";
 import { dashboardInitialState } from "src/core-logic/domain/admin/dashboardUrls/dashboardUrls.slice";
 import { fetchUserInitialState } from "src/core-logic/domain/admin/fetchUser/fetchUser.slice";
 import { listUsersInitialState } from "src/core-logic/domain/admin/listUsers/listUsers.slice";
 import { notificationsInitialState } from "src/core-logic/domain/admin/notifications/notificationsSlice";
+import { agencyAdminInitialState } from "src/core-logic/domain/agencies/fetch-agency-options/fetchAgencyOptions.slice";
 import type { RootState } from "src/core-logic/storeConfig/store";
 import { apiConsumerInitialState } from "../apiConsumer/apiConsumer.slice";
