+++ front/src/core-logic/domain/admin/connectedUsersAdmin/connectedUsersAdmin.test.ts
@@ -23,8 +23,8 @@ import {
   type TestDependencies,
 } from "src/core-logic/storeConfig/createTestStore";
 import type { ReduxStore } from "src/core-logic/storeConfig/store";
+import { agencyAdminInitialState } from "../../agencies/fetch-agency-options/fetchAgencyOptions.slice";
 import { feedbacksSelectors } from "../../feedback/feedback.selectors";
-import { agencyAdminInitialState } from "../agenciesAdmin/fetch-agency-options/fetchAgencyOptions.slice";
 import {
   type ConnectedUsersAdminFeedback,
   type ConnectedUsersAdminState,
